refactor: S29.03 extract shared FreeRTOS BDD-target pipeline
The DRY pass S29.01 deferred. The two FreeRTOS target main.c files
(Plus-TCP + lwIP) shared ~600 lines of near-verbatim code: the store /
security-policy lifecycle, the OnSet handler, the SD set, the Service
drain task, the console glue, ErrorHandlerEx. Extract all of it into
Bdd/Targets/Common/BddTargetFreeRtosPipeline.{c,h}; each main.c now keeps
only its network backend behind an injection seam.

The seam (BddTargetFreeRtosPipelineConfig): DefaultHost, a BuildSender
thunk (platform network bring-up + Switching sender, runs on the
interactive task so lwIP's adapters touch a started core), a GetHostname
callback, and a TeardownNetwork thunk. The pipeline owns the SolidSyslog
lifecycle / store / SD / both tasks and exports GetEndpoint /
GetEndpointVersion / Sleep / Exit / InitConsole. The Service task
self-registers its handle.

FreeRtos/main.c 1258 -> 291, FreeRtosLwip/main.c 1111 -> 320; the shared
surface (956 lines incl. header) now exists once (net -800 lines).

Pure refactor, no behaviour change. Both targets cross-build clean on
sha-0b93766; lwIP ELF has zero Platform/PlusTcp symbols; clang-format
clean; full BDD suite green on both targets (44/44 each: UDP/TCP/TLS/mTLS
+ the full @store suite incl. @hmac/@aesgcm/capacity/power-cycle).
